cleanup onnx legalizer
